[Cli] Move Exiter class from Support component to Cli/Server.
