+++ src/Valkyrja/Cli/Server/Handler/InputHandler.php
@@ -34,8 +34,8 @@ use Valkyrja\Cli\Routing\Data\Option\SilentOptionParameter;
 use Valkyrja\Cli\Routing\Data\Option\VersionOptionParameter;
 use Valkyrja\Cli\Routing\Dispatcher\Contract\Router;
 use Valkyrja\Cli\Server\Handler\Contract\InputHandler as Contract;
+use Valkyrja\Cli\Server\Support\Exiter;
 use Valkyrja\Container\Manager\Contract\Container;
-use Valkyrja\Support\Exiter;
 
 /**
  * Class InputHandler.
