+++ tests/Unit/Application/Entry/AppTest.php
@@ -24,6 +24,7 @@ use Valkyrja\Cli\Interaction\Output\Output;
 use Valkyrja\Cli\Routing\Collection\Contract\Collection as CliCollection;
 use Valkyrja\Cli\Routing\Data\Data as CliData;
 use Valkyrja\Cli\Routing\Data\Route;
+use Valkyrja\Cli\Server\Support\Exiter;
 use Valkyrja\Container\Data\Data as ContainerData;
 use Valkyrja\Dispatch\Data\MethodDispatch;
 use Valkyrja\Event\Data\Data as EventData;
@@ -32,7 +33,6 @@ use Valkyrja\Http\Routing\Collection\Contract\Collection as HttpCollection;
 use Valkyrja\Http\Routing\Data\Data as HttpData;
 use Valkyrja\Http\Routing\Data\Route as HttpRoute;
 use Valkyrja\Support\Directory;
-use Valkyrja\Support\Exiter;
 use Valkyrja\Support\Microtime;
 use Valkyrja\Tests\EnvClass;
 use Valkyrja\Tests\Unit\TestCase;
